fix: repair pairing handoff, close secret-validation hole, harden admin roles
Fixes ten issues found auditing the previous security-remediation pass, plus
that pass's own uncommitted work. Full detail in cloud-screen-audit-v2.md.

Critical
- Pairing issued a per-screen secret and both `player/bootstrap` and the
  `screen:register` gateway message began requiring it, but the player app
  never received, stored, or sent it. Every newly paired screen 401'd. It
  escaped review because seeded/demo screens are created outside the pairing
  flow, so they have a NULL hash and still work via the shared-secret
  fallback -- those were the only screens exercised.
  The poll response now carries `screenSecret`; the player persists it before
  reloading; the `pairing:complete` socket broadcast only prompts a poll,
  since a room broadcast can never carry a credential.
- assertProductionSecretsAreSet() blocklisted only the three `dev-*` source
  fallbacks, so docker-compose's own `change-me-in-production` default passed
  the check -- a publicly committed JWT signing key reported as a clean boot.
  Replaced with length (>=32), character-variety, and access!=refresh rules;
  the blocklist is now the second line of defence, not the first.
  Tokens also carry a `typ` claim so a refresh token cannot be replayed as a
  Bearer access token even if both keys are misconfigured to the same value.
- docker-compose set NODE_ENV=production alongside a PLAYER_HEARTBEAT_SECRET
  default that the boot check rejects, so `docker compose up` could not start.
  The dangerous defaults are gone; Compose now fails fast on a missing secret.

High
- PlatformStaffDbGuard admitted any staff role to every admin route not
  individually re-locked, including the cross-tenant audit log. Rewritten
  fail-closed behind a @PlatformRoles decorator: a route naming no roles is
  super-admin-only, so omitting the decorator locks a route rather than
  opening it. Writes stay super-admin-only pending a product decision.
- media: a file of up to MAX_BYTES was written inside a Prisma interactive
  transaction holding pg_advisory_xact_lock -- past the 5s default timeout on
  large uploads, and not rolled back on failure. Bytes now land in a temp file
  before the transaction; an atomic rename follows the commit. The DB row is
  removed if the rename fails, so the quota never counts an absent file.
- duplicateMediaToWorkspace skipped the storage quota entirely; both paths now
  share assertWithinStorageQuotaTx().
- The audit trail was a JSON file appended with an unlocked read-modify-write,
  so concurrent impersonation events dropped each other, and it sat outside
  every database backup. Moved to an AuditLog table.
- scripts/backup-uploads.sh archived only the media volume, leaving the
  database unprotected. Replaced by scripts/backup.sh (pg_dump + both volumes)
  and scripts/restore.sh.
- Dockerfile.backend exposed and probed :4000 while the process listens on
  PORT||3000, so the container was permanently unhealthy outside compose.
- AllExceptionsFilter is registered via APP_FILTER and so also received
  WebSocket exceptions, where switchToHttp().getResponse() returns the message
  payload and .status() threw from inside the filter, masking the real error.
  Now dispatches on host.getType().

Also: `.part` staging files are never served; sizeBytes comes from the buffer
rather than the client; the long-standing TS7022 in stripe-webhook.service.spec
is fixed.

Test infrastructure
- `await import('file-type')` (pure ESM) could not execute under Jest, leaving
  the magic-byte upload check untestable. The backend test script now runs with
  --experimental-vm-modules, so that path is covered for real.
- apps/player had no test runner; added one via `tsx --test` (no new deps).

Backend: 10 suites / 53 tests (was 8 / 21). Player: 6 tests (was none).
tsc --noEmit, eslint, nest build, next build, and i18n:check all clean.

+++ apps/backend/src/domains/workspaces/workspaces.service.ts
@@ -56,7 +56,7 @@ export class WorkspacesService {
               status: SubscriptionStatus.TRIALING,
               seats: 5,
               screenLimit: 25,
-              storageLimitBytes: 5 * 1024 * 1024 * 1024,
+              storageLimitBytes: BigInt(5 * 1024 * 1024 * 1024),
             },
           },
         },
